rename: utils 패키지 ->  validator, message, generator 패키지로 구분
