+++ src/main/java/racingcar/utils/message/GameMessage.java
@@ -1,4 +1,4 @@
-package racingcar.utils;
+package racingcar.utils.message;
 
 public enum GameMessage {
     REQUEST_CAR_NAMES_MESSAGE("경주할 자동차 이름을 입력하세요.(이름은 쉼표(,) 기준으로 구분)"),
